requirements.txt
python-dotenv
openai
pypdf
requests
